LIME-81: * exchangeToken
